Eclipse2024 - ShirinZaidi
I heard you wanted an eclipse so I got you an eclipse in front of the eclipsed eclipse [The sky in Rochester, NY, circa 2024. The image is of a disk made of obsidian? Some mysterious otherworldly element. Something to evoke the haunting nature of witnessing a ~4m dark sky during an eclipse event. In front of a fully shrouded sky, minutes before the actual eclipse and rainfall begins. Proof that no matter how high of a grade you get in your undergraduate meteorology course, you will never be able to get close to predicting the best sky to observe under.]