Schematic - ShirinZaidi
This is a schematic I created for my talk on Thursday (tomorrow). It shows our Boss & Bodenheimer 1979 dense core collapse setup, which has a thermal-to-grav energy ratio of 0.4 (so a ~0.02 pc cloud w/ a density of ~2.3e-18 g/cm^3 unstable enough to collapse), a rotational-to-grav energy ratio of 0.04, and a mass-to-flux ratio of ~3.3. We start with this cloud and within a free-fall timescale, see a dense solar mass core embedded in a gas & dust disk that expands from 20 au to 40 au in the span of ~30 kyr, due to accretion.